bin filter

+++ source/web_server/phenotypePredictionApp/static/phenotypePredictionApp/result_figures.js
@@ -77,7 +77,7 @@
         function __initialize_verdict_cutoff_spinner(dataTable) {
             $.fn.dataTable.ext.search.push(
                 function( settings, data, dataIndex ) {
-                    var min = parseInt( $('#dt_results_verdict_filter').val(), 10 );
+                    var min = parseFloat( $('#dt_results_verdict_filter').val(), 10 );
                     var verdict = parseFloat( data[3] ) || 0;
 
                     if  ( isNaN( min ) ||
